+++ code/matrix/src/matrix_1darray.c
@@ -9,14 +9,14 @@
 #include "err.h"
 
 struct _matrix {
-	int* rows;
+	double* rows;
 	int num_rows;
 	int num_cols;
 };
 
 matrix* matrix_new(int num_rows, int num_cols) {
 	matrix* m = malloc(sizeof(matrix));
-	m->rows = malloc(sizeof(int) * num_rows * num_cols);
+	m->rows = malloc(sizeof(double) * num_rows * num_cols);
 	m->num_rows = num_rows;
 	m->num_cols = num_cols;
 
@@ -40,15 +40,15 @@ void matrix_free(matrix* m) {
 	free(m);
 }
 
-int* matrix_at(matrix* m, int row, int col) {
+double* matrix_at(matrix* m, int row, int col) {
 	return &(m->rows[row * matrix_num_cols(m) + col]);
 }
 
-void matrix_set(matrix* m, int row, int col, int val) {
+void matrix_set(matrix* m, int row, int col, double val) {
 	*matrix_at(m, row, col) = val;
 }
 
-void matrix_smul(matrix* m, int x) {
+void matrix_smul(matrix* m, double x) {
 	int i, j;
 	for (i = 0; i < matrix_num_rows(m); i++) {
 		for (j = 0; j < matrix_num_cols(m); j++) {
@@ -57,7 +57,7 @@ void matrix_smul(matrix* m, int x) {
 	}
 }
 
-int matrix_get(matrix* m, int row, int col) {
+double matrix_get(matrix* m, int row, int col) {
 	return *matrix_at(m, row, col);
 }
